7번 문제 통과

+++ inflearn/lecture06/problem04/Main.java
@@ -24,7 +24,7 @@ public class Main {
 
         for (int n : arr) {
             if (cache.contains(n)) {
-                cache.remove(n);
+                cache.remove((Integer) n);
             }
             cache.addFirst(n);
